requirements.txt
# 主依赖 (运行时)
# cflib 0.1.x API 在小版本间稳定；0.2 之前锁定以避免 link_statistics Caller 接口变动。
cflib>=0.1.30,<0.2
numpy>=1.24,<3
scipy>=1.10
matplotlib>=3.7
PyYAML>=6
cvxpy>=1.4

# Web 前端 (PR-A: 离线分析 Web UI)
fastapi>=0.110
uvicorn[standard]>=0.27
jinja2>=3.1

# 可选: C++ radio driver 后端 (需要 MSVC Build Tools)
# cflinkcpp
